# internal/docs/swagger.yaml
basePath: /api
definitions:
...
  projects.createProjectRequest:
    properties:
      description:
        type: string
      name:
        maxLength: 160
        minLength: 2
        type: string
    required:
    - name
    type: object
  projects.inviteMemberRequest:
    properties:
      email:
        type: string
      expires_at:
        type: string
      role:
        type: string
    required:
    - email
    - expires_at
    - role
    type: object
  projects.projectInvitationResponse:
    properties:
      created_at:
        type: string
      email:
        type: string
      expires_at:
        type: string
      id:
        type: string
      project_id:
        type: string
      role:
        type: string
      status:
        type: string
    type: object
  projects.projectMemberResponse:
    properties:
      created_at:
        type: string
      id:
        type: string
      project_id:
        type: string
      role:
        type: string
      status:
        type: string
      updated_at:
        type: string
      user_id:
        type: string
    type: object
  projects.projectResponse:
    properties:
      created_at:
        type: string
      description:
        type: string
      id:
        type: string
      name:
        type: string
      owner_id:
        type: string
      status:
        type: string
      updated_at:
        type: string
    type: object
  projects.updateProjectRequest:
    properties:
      description:
        type: string
      name:
        type: string
      status:
        type: string
    type: object
...
host: localhost:8080
info:
  contact: {}
  description: API with Clean Architecture
  title: Indagio API
  version: "1.0"
paths:
...
  /projects:
    get:
      description: List projects where the authenticated user is owner or member
      parameters:
      - description: Filter by project status
        in: query
        name: status
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/projects.projectResponse'
            type: array
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: List projects
      tags:
      - Projects
    post:
      consumes:
      - application/json
      description: Create a new project owned by the authenticated user
      parameters:
      - description: Project payload
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/projects.createProjectRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/projects.projectResponse'
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "422":
          description: Unprocessable Entity
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Create project
      tags:
      - Projects
  /projects/{id}:
    delete:
      description: Delete a project permanently
      parameters:
      - description: Project ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "204":
          description: No Content
          schema:
            type: string
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Delete project
      tags:
      - Projects
    get:
      description: Get one project by ID when the authenticated user has access
      parameters:
      - description: Project ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/projects.projectResponse'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Get project by ID
      tags:
      - Projects
    patch:
      consumes:
      - application/json
      description: Update project name, description, or status
      parameters:
      - description: Project ID
        in: path
        name: id
        required: true
        type: string
      - description: Project patch payload
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/projects.updateProjectRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/projects.projectResponse'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "422":
          description: Unprocessable Entity
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Update project
      tags:
      - Projects
...
  /projects/{id}/archive:
    post:
      description: Archive a project and make it read-only for normal operations
      parameters:
      - description: Project ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Archive project
      tags:
      - Projects
...
  /projects/{id}/members:
    get:
      description: List all members in a project
      parameters:
      - description: Project ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/projects.projectMemberResponse'
            type: array
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: List project members
      tags:
      - Projects
  /projects/{id}/members/invite:
    post:
      consumes:
      - application/json
      description: Invite a user by email to join a project with a role
      parameters:
      - description: Project ID
        in: path
        name: id
        required: true
        type: string
      - description: Invitation payload
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/projects.inviteMemberRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/projects.projectInvitationResponse'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "422":
          description: Unprocessable Entity
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Invite project member
      tags:
      - Projects
...
schemes:
- http
- https
securityDefinitions:
  BearerAuth:
    description: Type "Bearer" followed by space and JWT token.
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"